Compute some statistics from input
This little programm reads from input and depending from the parameters
it computes the min, max, mean, variance, standard deviation or median.

Thereby you can ignore the first or last line. This is for example useful,
when you give the output of wc -l into this programm. wc -l has a last line
which sums up all lines in the listed files.

Furthermore there is the gentle flag which gives you control if you want to
exit on errors (e.g. input not a number or overflow). In oth cases - if the
gentle flag is not set - the programm proceeds without these values.

Signed-off-by: bhbb <bhbboerger@gmail.com>