main.cpp: add option --weigthed
In the internal stats was from the beginning capable to compute weithed means.
But it was not possible to feed in weigths, because stats interprets all input
as data points.
By adding the option --weigthed stats treats every second number as weigth of
the following number. The first number is treated as weight.

Example:
	Input:
	2, 20, 100, 2

	Output of stats --mean --weigthed:
	2.35294

Signed-off-by: bhbb <bhbboerger@gmail.com>