stats.h: fix memory leak and heap overflow
When allocating memory for the weights we have to free it after we called
compute_weigthed_mean with the weigths.

There was a heap overflow because we allocated length bytes of memory.
This assumes that integer is exactly one byte long. This assumption is wrong.
After the standard an integer have to be at least 2 bytes. Fix this by
multiplying this with the size of an integer.

Signed-off-by: bhbb <bhbboerger@gmail.com>